+++ Project_Spring/src/main/java/com/spring/project/repository/RoleRepsitory.java
@@ -4,7 +4,7 @@ import com.spring.project.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RoleRepsitory extends JpaRepository<Role, Integer> {
+public interface RoleRepsitory extends JpaRepository<Role, Long> {
 
     Role findByName(String name);
 }
