+++ Project_Spring/src/main/java/com/spring/project/repository/SubscriptionHistoryRepository.java
@@ -10,20 +10,20 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface SubscriptionHistoryRepository extends JpaRepository<SubscriptionsHistory, Integer> {
+public interface SubscriptionHistoryRepository extends JpaRepository<SubscriptionsHistory, Long> {
     List<SubscriptionsHistory> findAll();
     @Query("SELECT sh FROM SubscriptionsHistory sh " +
             "WHERE sh.user.id = :userId " +
             "ORDER BY sh.subscriptionEndTime DESC")
-    List<SubscriptionsHistory> findByUser_IdOrderBySubscriptionEndTimeAsc(@Param("userId") Integer userId);
+    List<SubscriptionsHistory> findByUser_IdOrderBySubscriptionEndTimeAsc(@Param("userId") Long userId);
 
     @Query("SELECT sh FROM SubscriptionsHistory sh " +
             "WHERE sh.user.id = :userId " +
             "AND sh.subscriptionEndTime >= :date " +
             "ORDER BY sh.subscriptionEndTime DESC")
-    SubscriptionsHistory findActiveSubscriptionForUser(@Param("userId") Integer userId, @Param("date") LocalDate date);
+    SubscriptionsHistory findActiveSubscriptionForUser(@Param("userId") Long userId, @Param("date") LocalDate date);
 
     @Transactional
     @Modifying
-    void deleteAllByUser_Id(@Param("userId") Integer userId);
+    void deleteAllByUser_Id(@Param("userId") Long userId);
 }
