+++ Project_Spring/src/main/java/com/spring/project/service/FotballReservationService.java
@@ -8,7 +8,7 @@ public interface FotballReservationService {
 
     Reservation save(Reservation reservation);
 
-    void deleteReservation(Integer id);
+    void deleteReservation(Long id);
 
     List<Reservation> getReservationsByCourt(String court);
