+++ Project_Spring/src/main/java/com/spring/project/service/impl/FotballReservationServiceImpl.java
@@ -20,7 +20,7 @@ public class FotballReservationServiceImpl implements FotballReservationService 
     }
 
     @Override
-    public void deleteReservation(Integer id){
+    public void deleteReservation(Long id){
         reservationRepository.deleteById(id);
     }
